Reverting recent changes because they broke svnmerge.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/release-4_2_2@29693 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/display/Display/PSPixelCanvas.cc
@@ -670,9 +670,9 @@ namespace casa { //# NAMESPACE CASA - BEGIN
 					itsComponents[i+2] = (uShort)(b*maxcolor3);
 				}
 			}
-			drawPSImage(static_cast<int>(itsComponentX), static_cast<int>(itsComponentY),
-					itsComponentWidth, itsComponentHeight, 1.0, 1.0,
-			        itsComponents, itsComponentBPC, PSpcctbl()->colorModel());
+			drawPSImage(itsComponentX, itsComponentY, itsComponentWidth, itsComponentHeight,
+			            1.0, 1.0,
+			            itsComponents, itsComponentBPC, PSpcctbl()->colorModel());
 			itsComponentWidth = itsComponentHeight = 0;
 			delete [] itsComponents;
 			itsComponents = NULL;
